ruoyi-fastapi-backend/requirements.txt
agno==2.4.8
aiofiles==25.1.0
alembic==1.18.3
anthropic==0.78.0
APScheduler==3.11.2
async-lru==2.1.0
asyncmy==0.2.11
bcrypt==5.0.0
cerebras-cloud-sdk==1.67.0
cohere==5.20.4
fastapi[all]==0.128.2
google-genai==1.62.0
groq==1.0.0
litellm==1.81.8
llama-index-core==0.14.19
llama-index-vector-stores-milvus==1.1.0
llama-api-client==0.6.0
loguru==0.7.3
mistralai==1.12.0
ollama==0.6.1
openai==2.17.0
openpyxl==3.1.5
pandas==2.3.3
Pillow==12.1.1
portkey-ai==2.1.0
psutil==7.2.2
pymilvus==2.6.7
pypdf==6.1.1
pydantic>=2.11.4
pydantic-settings==2.13.1
pydantic-validation-decorator==0.1.5
PyJWT[crypto]==2.11.0
PyMySQL==1.1.2
python-dotenv==1.2.2
python-docx==1.2.0
redis==7.1.0
ruff==0.15.0
SQLAlchemy[asyncio]==2.0.46
sqlglot[rs]==28.10.0
user-agents==2.2.0
